rust: fix clippy lints for clippy::unnecessary_cast

+++ rust/src/dns/dns.rs
@@ -507,7 +507,7 @@ impl DNSState {
         let mut consumed = 0;
         while !cur_i.is_empty() {
             if cur_i.len() == 1 {
-                return AppLayerResult::incomplete(consumed as u32, 2 as u32);
+                return AppLayerResult::incomplete(consumed as u32, 2_u32);
             }
             let size = match be_u16(cur_i) as IResult<&[u8],u16> {
                 Ok((_, len)) => len,
@@ -556,7 +556,7 @@ impl DNSState {
         let mut consumed = 0;
         while !cur_i.is_empty() {
             if cur_i.len() == 1 {
-                return AppLayerResult::incomplete(consumed as u32, 2 as u32);
+                return AppLayerResult::incomplete(consumed as u32, 2_u32);
             }
             let size = match be_u16(cur_i) as IResult<&[u8],u16> {
                 Ok((_, len)) => len,
